LayoutTests/fast/multicol/zeroColumnCount.html

<!DOCTYPE html>
<body>
<style>
body {
	-webkit-column-count: 2;
	-webkit-column-count: 0;
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
<script>
description("Test that column-count:0 is parsed as invalid syntax and thrown away.");
shouldBe("getComputedStyle(document.body).webkitColumnCount", "'2'");
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<script src="../js/resources/js-test-post.js"></script>
