third_party/blink/web_tests/platform/generic/fast/multicol/zeroColumnCount-expected.txt

Test that column-count:0 is parsed as invalid syntax and thrown away.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.body).webkitColumnCount is '2'
PASS successfullyParsed is true

TEST COMPLETE
